+++ client/src/pages/Home.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Row  from "react-bootstrap/Row";
 import Navigation from "../components/Navigation";
 import Slides from "../components/Slides";
@@ -7,47 +7,10 @@ import Jumbotron from '../components/Jumbotron';
 import Button from 'react-bootstrap/Button';
 import {Link} from "react-router-dom";
 import "./style.css";
-import axios from "axios";
-import { useStoreContext } from "../utils/GlobalState";
-import { SET_CURRENT_USER } from "../utils/actions";
+
 
 
 function Home() {
-  const [state, dispatch] = useStoreContext();
-  const [user, setUser] = useState();
-
-  useEffect(() => {
-    const checkLoggedIn = async () => {
-      let token = localStorage.getItem("auth-token");
-      if (token === null) {
-        localStorage.setItem("auth-token", "");
-        token = "";
-      }
-      const tokenRes = await axios.post(
-        "/api/user/validate",
-        null,
-        { headers: { "x-auth-token": token } }
-      );
-      console.log(tokenRes.data);
-      if (tokenRes.data) {
-        const userRes = await axios.get(
-          "/api/user",
-          { headers: { "x-auth-token": token } }
-        );
-        dispatch({
-          type: SET_CURRENT_USER,
-          user: userRes.data
-        });
-        setUser({
-          token,
-          user: userRes.data,
-        });
-      }
-    };
-
-    checkLoggedIn();
-  }, []);
-
     return (
 
         <div>
