+++ client/src/pages/Submit.js
@@ -1,52 +1,12 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Jumbotron from "react-bootstrap/Jumbotron";
 import Navigation from "../components/Navigation";
 import SubmitForm from "../components/SubmitForm";
 import Header from "../components/Header";
-import axios from "axios";
-import { useHistory } from "react-router-dom";
-import { useStoreContext } from "../utils/GlobalState";
-import { SET_CURRENT_USER } from "../utils/actions";
+
 
 function Submit() {
-  const [state, dispatch] = useStoreContext();
-  const [user, setUser] = useState();
-  const history = useHistory();
-
-  useEffect(() => {
-    const checkLoggedIn = async () => {
-      let token = localStorage.getItem("auth-token");
-      if (token === null) {
-        localStorage.setItem("auth-token", "");
-        token = "";
-      }
-      const tokenRes = await axios.post(
-        "/api/user/validate",
-        null,
-        { headers: { "x-auth-token": token } }
-      );
-      console.log(tokenRes.data);
-      if (tokenRes.data) {
-        const userRes = await axios.get(
-          "/api/user",
-          { headers: { "x-auth-token": token } }
-        );
-        dispatch({
-          type: SET_CURRENT_USER,
-          user: userRes.data
-        });
-        setUser({
-          token,
-          user: userRes.data,
-        });
-      } else {
-        return (history.push("/signin"));
-      }
-    };
-
-    checkLoggedIn();
-  }, []);
-
+  if (state.currentUser._id !== 0) {
     return (
       <div>
         <div>
@@ -59,7 +19,15 @@ function Submit() {
         </div>
       </div>
     )
-  
+  } else {
+    return (
+      <div>
+        <Header />
+        <Navigation />
+        <SignIn />
+      </div>
+    )
+  }
 
 }
